Move buche code to own subdirectory
